requirements.txt
# Core Dependencies
numpy>=1.26.0
pandas>=2.0.0
matplotlib>=3.8.0
Pillow>=10.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
safetensors>=0.4.0

# Web App
streamlit>=1.28.0

# ML Utilities
scikit-learn>=1.3.0

# General Utilities
requests>=2.31.0
python-dotenv>=1.0.0
